Refactor Duke into packages
