--- src/main/java/Deadline.java
@@ -1,47 +1,0 @@
-import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
-
-public class Deadline extends Task{
-
-    private LocalDateTime date;
-
-    public Deadline(String taskName) throws InvalidTaskDescriptionException,InvalidTaskDataTimeException{
-        super();
-        String[] str =  taskName.split("/by");
-
-        try {
-            this.taskName = str[0].strip();
-            if (this.taskName.equals("")) this.taskName=null;
-        } catch (IndexOutOfBoundsException i){
-            this.taskName=null;
-        }
-        try {
-            this.date = Parser.parseDateTime(str[1].strip());
-        } catch (IndexOutOfBoundsException i){
-            this.date=null;
-        }
-
-        if (this.taskName == null){
-            throw new InvalidTaskDescriptionException("Deadline");
-        }
-        if (this.date == null){
-            throw new InvalidTaskDataTimeException("Deadline");
-        }
-    }
-
-    public Deadline(String taskName, LocalDateTime date){
-        this.taskName = taskName;
-        this.date = date;
-    }
-
-    public char getType(){
-        return 'D';
-    }
-
-    @Override
-    public String toString(){
-        DateTimeFormatter formatted = DateTimeFormatter.ofPattern("d MMM yyyy, K:mma");
-
-        return String.format("[%c][%c] %s (by: %s)",this.getType(),this.done,this.taskName,this.date.format(formatted));
-    }
-}
