--- src/main/java/DeleteCommand.java
@@ -1,9 +1,0 @@
-public class DeleteCommand extends Command{
-    public DeleteCommand(String userInput){
-        super(userInput);
-    }
-
-    public boolean execute(Storage storage, Ui ui, TaskManager taskManager){
-        return false;
-    }
-}
