--- src/main/java/ExitCommand.java
@@ -1,15 +1,0 @@
-public class ExitCommand extends Command{
-    public ExitCommand(){
-        super();
-    }
-
-    @Override
-    public boolean isExit(){
-        return true;
-    }
-
-    public boolean execute(Storage storage, Ui ui, TaskManager taskManager){
-        ui.showBye();
-        return true;
-    }
-}
