--- src/main/java/MarkCommand.java
@@ -1,15 +1,0 @@
-public abstract class MarkCommand extends Command{
-    public MarkCommand(String userInput){
-        super(userInput);
-    }
-
-    public static MarkCommand of(String userInput){
-        if (userInput.startsWith("mark")) {
-            return new MarkDoneCommand(userInput);
-        } else {
-            return new MarkUndoneCommand(userInput);
-        }
-    }
-
-    public abstract boolean execute(Storage storage, Ui ui, TaskManager taskManager) throws DukeException;
-}
