+++ src/main/java/duke/commands/Command.java
@@ -1,0 +1,33 @@
+package duke.commands;
+
+import duke.exceptions.DukeException;
+import duke.Storage;
+import duke.TaskManager;
+import duke.Ui;
+
+public abstract class Command {
+    Storage storage = new Storage();
+    Ui ui = new Ui();
+    TaskManager taskManager = new TaskManager();
+    String userInput = "";
+
+    public Command(String userInput){
+        this.userInput = userInput;
+    }
+    public Command(){}
+
+    protected void save(Storage storage,Ui ui, TaskManager taskManager){
+        try {
+            storage.saveTaskManager(taskManager);
+            ui.showSavingComplete();
+        } catch (DukeException e){
+            ui.showSavingFailed();
+        }
+    }
+
+    public abstract boolean execute(Storage storage, Ui ui, TaskManager taskManager) throws DukeException;
+
+    public boolean isExit(){
+        return false;
+    }
+}
