+++ src/main/java/duke/commands/ListCommand.java
@@ -1,0 +1,16 @@
+package duke.commands;
+
+import duke.Storage;
+import duke.TaskManager;
+import duke.Ui;
+
+public class ListCommand extends Command{
+    public ListCommand(){
+        super();
+    }
+
+    public boolean execute(Storage storage, Ui ui, TaskManager taskManager){
+        ui.showList(taskManager);
+        return true;
+    }
+}
