+++ src/main/java/duke/exceptions/FailedTaskParseException.java
@@ -1,4 +1,6 @@
-public class FailedTaskParseException extends DukeException{
+package duke.exceptions;
+
+public class FailedTaskParseException extends DukeException {
     public FailedTaskParseException(String userInput){
         super("Failed to parse task from: " + userInput);
     }
