+++ src/main/java/duke/tasks/Todo.java
@@ -1,0 +1,14 @@
+package duke.tasks;
+
+import duke.exceptions.DukeException;
+
+public class Todo extends Task{
+
+    public Todo(String taskName){
+        this.taskName = taskName;
+    }
+
+    public char getType(){
+        return 'T';
+    }
+}
