+++ src/test/java/DeadlineTest.java
@@ -1,3 +1,4 @@
+import duke.tasks.Deadline;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
