+++ src/test/java/EventTest.java
@@ -1,3 +1,4 @@
+import duke.tasks.Event;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.BeforeEach;
