+++ src/test/java/TodoTest.java
@@ -1,3 +1,4 @@
+import duke.tasks.Todo;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
